Moves bundle_profile to segment.bundles and renames it.
Also, multiple other changes according to review by @garyfallidis.
