+++ dipy/align/streamlinear.py
@@ -290,14 +290,15 @@ class StreamlineLinearRegistration(object):
         References
         ----------
         .. [Garyfallidis15] Garyfallidis et al. "Robust and efficient linear
-                            registration of white-matter fascicles in the space
-                            of streamlines", NeuroImage, 117, 124--140, 2015
+           registration of white-matter fascicles in the space
+           of streamlines", NeuroImage, 117, 124--140, 2015
+
         .. [Garyfallidis14] Garyfallidis et al., "Direct native-space fiber
-                            bundle alignment for group comparisons", ISMRM,
-                            2014.
+           bundle alignment for group comparisons", ISMRM, 2014.
+
         .. [Garyfallidis17] Garyfallidis et al. Recognition of white matter
-                            bundles using local and global streamline-based
-                            registration and clustering, Neuroimage, 2017.
+           bundles using local and global streamline-based
+           registration and clustering, Neuroimage, 2017.
         """
 
         self.x0 = self._set_x0(x0)
