+++ dipy/segment/bundles.py
@@ -1,6 +1,9 @@
 import numpy as np
+from scipy.spatial.distance import mahalanobis
 from dipy.tracking.streamline import (set_number_of_points, nbytes,
-                                      select_random_set_of_streamlines)
+                                      select_random_set_of_streamlines,
+                                      values_from_volume,
+                                      orient_by_streamline)
 from dipy.segment.clustering import qbx_and_merge
 from dipy.tracking.distances import (bundles_distances_mdf,
                                      bundles_distances_mam)
@@ -632,3 +635,153 @@ class RecoBundles(object):
             print(' Duration %0.3f sec. \n' % (time() - t, ))
 
         return pruned_streamlines, labels
+
+
+def gaussian_weights(bundle, n_points=100, return_mahalnobis=False,
+                     stat=np.mean):
+    """
+    Calculate weights for each streamline/node in a bundle, based on a
+    Mahalanobis distance from the core the bundle, at that node (mean, per
+    default).
+
+    Parameters
+    ----------
+    bundle : Streamlines
+        The streamlines to weight.
+    n_points : int, optional
+        The number of points to resample to. *If the `bundle` is an array, this
+        input is ignored*. Default: 100.
+
+    Returns
+    -------
+    w : array of shape (n_streamlines, n_points)
+        Weights for each node in each streamline, calculated as its relative
+        inverse of the Mahalanobis distance, relative to the distribution of
+        coordinates at that node position across streamlines.
+    """
+    # Resample to same length for each streamline:
+    bundle = set_number_of_points(bundle, n_points)
+
+    # This is the output
+    w = np.zeros((len(bundle), n_points))
+
+    # If there's only one fiber here, it gets the entire weighting:
+    if len(bundle) == 1:
+        if return_mahalnobis:
+            return np.array([np.nan])
+        else:
+            return np.array([1])
+
+    for node in range(n_points):
+        # This should come back as a 3D covariance matrix with the spatial
+        # variance covariance of this node across the different streamlines
+        # This is a 3-by-3 array:
+        node_coords = bundle.data[node::n_points]
+        c = np.cov(node_coords.T, ddof=0)
+        # Reorganize as an upper diagonal matrix for expected Mahalnobis input:
+        c = np.array([[c[0, 0], c[0, 1], c[0, 2]],
+                      [0, c[1, 1], c[1, 2]],
+                      [0, 0, c[2, 2]]])
+        # Calculate the mean or median of this node as well
+        # delta = node_coords - np.mean(node_coords, 0)
+        m = stat(node_coords, 0)
+        # Weights are the inverse of the Mahalanobis distance
+        for fn in range(len(bundle)):
+            # In the special case where all the streamlines have the exact same
+            # coordinate in this node, the covariance matrix is all zeros, so
+            # we can't calculate the Mahalnobis distance, we will instead give
+            # each streamline an identical weight, equal to the number of
+            # streamlines:
+            if np.allclose(c, 0):
+                w[:, node] = len(bundle)
+                break
+            # Otherwise, go ahead and calculate Mahalanobis for node on
+            # fiber[fn]:
+            w[fn, node] = mahalanobis(node_coords[fn], m, np.linalg.inv(c))
+    if return_mahalnobis:
+        return w
+    # weighting is inverse to the distance (the further you are, the less you
+    # should be weighted)
+    w = 1 / w
+    # Normalize before returning, so that the weights in each node sum to 1:
+    return w / np.sum(w, 0)
+
+
+def afq_tract_profile(data, bundle, affine=None, n_points=100,
+                      orient_by=None, weights=None, **weights_kwarg):
+    """
+    Calculates a summarized profile of data for a bundle or tract
+    along its length.
+
+    Follows the approach outlined in [Yeatman2012]_.
+
+    Parameters
+    ----------
+    data : 3D volume
+        The statistic to sample with the streamlines.
+
+    bundle : StreamLines class instance
+        The collection of streamlines (possibly already resampled into an array
+         for each to have the same length) with which we are resampling. See
+         Note below about orienting the streamlines.
+
+    affine: 4-by-4 array, optional.
+        A transformation associated with the streamlines in the bundle.
+        Default: identity.
+
+    n_points: int, optional
+        The number of points to sample along the bundle. Default: 100.
+
+    orient_by: streamline, optional.
+        A streamline to use as a standard to orient all of the streamlines in
+        the bundle according to.
+
+    weights : 1D array or 2D array or callable (optional)
+        Weight each streamline (1D) or each node (2D) when calculating the
+        tract-profiles. Must sum to 1 across streamlines (in each node if
+        relevant). If callable, this is a function that calculates weights.
+
+    weights_kwarg : key-word arguments
+        Additional key-word arguments to pass to the weight-calculating
+        function. Only to be used if weights is a callable.
+
+    Returns
+    -------
+    ndarray : a 1D array with the profile of `data` along the length of
+        `bundle`
+
+    Note
+    ----
+    Before providing a bundle as input to this function, you will need to make
+    sure that the streamlines in the bundle are all oriented in the same
+    orientation relative to the bundle (use :func:`orient_by_streamline`).
+
+    References
+    ----------
+    .. [Yeatman2012] Yeatman, Jason D., Robert F. Dougherty,
+       Nathaniel J. Myall, Brian A. Wandell, and Heidi M. Feldman. 2012.
+       "Tract Profiles of White Matter Properties: Automating Fiber-Tract
+       Quantification" PloS One 7 (11): e49790.
+    """
+    if orient_by is not None:
+        bundle = orient_by_streamline(bundle, orient_by, affine=affine)
+    if len(bundle) == 0:
+        raise ValueError("The bundle contains no streamlines")
+
+    # Resample each streamline to the same number of points:
+    fgarray = set_number_of_points(bundle, n_points)
+
+    # Extract the values
+    values = np.array(values_from_volume(data, fgarray, affine=affine))
+
+    if weights is None:
+        weights = np.ones(values.shape) / values.shape[0]
+    elif callable(weights):
+        weights = weights(bundle, **weights_kwarg)
+    else:
+        # We check that weights *always sum to 1 across streamlines*:
+        if not np.allclose(np.sum(weights, 0), np.ones(n_points)):
+            raise ValueError("The sum of weights across streamlines must ",
+                             "be equal to 1")
+
+    return np.sum(weights * values, 0)
